feat: eye progress experiments
